requirements.txt
dj-database-url==0.4.2
Django==1.11.4
django-tastypie==0.14.0
gunicorn==19.7.1
psycopg2==2.7.3
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
six==1.10.0
